Ensure config file is always created if possible
