+++ UndercutF1.Console/CommandHandler.Root.cs
@@ -90,7 +90,7 @@ public static partial class CommandHandler
             app.MapControlEndpoints().MapTimingEndpoints();
         }
 
-        app.Logger.LogDebug("Options: {Options}", options);
+        app.Logger.LogInformation("{Options}", options);
 
         Whisper.net.Logger.LogProvider.AddLogger(
             (level, msg) =>
@@ -113,6 +113,40 @@ public static partial class CommandHandler
             }
         );
 
+        await EnsureConfigFileExistsAsync(app.Logger);
+
         await app.RunAsync();
+    }
+
+    private static async Task EnsureConfigFileExistsAsync(ILogger logger)
+    {
+        try
+        {
+            if (File.Exists(Options.ConfigFilePath))
+            {
+                return;
+            }
+            var schemaLocation =
+                "https://raw.githubusercontent.com/JustAman62/undercut-f1/refs/heads/master/config.schema.json";
+            var defaultConfigFile = $$"""
+                {
+                    "$schema": "{{schemaLocation}}"
+                }
+                """;
+
+            logger.LogInformation(
+                "Writing default configuration file to {Path}",
+                Options.ConfigFilePath
+            );
+            await File.WriteAllTextAsync(Options.ConfigFilePath, defaultConfigFile);
+        }
+        catch (Exception ex)
+        {
+            logger.LogError(
+                ex,
+                "Failed to write default configuration file to {Path}",
+                Options.ConfigFilePath
+            );
+        }
     }
 }
