+++ UndercutF1.Console/Options.cs
@@ -3,8 +3,23 @@ using UndercutF1.Data;
 
 namespace UndercutF1.Console;
 
+/// <summary>
+/// Options specific to the console app part of UndercutF1.
+/// </summary>
 public sealed record Options : LiveTimingOptions
 {
+    /// <summary>
+    /// Try to conform to Windows/XDG directory standards by default.
+    /// <see cref="Environment.SpecialFolder.ApplicationData"/> will return <c>%APPDATA%</c> on Windows,
+    /// <c>$XDG_CONFIG_HOME</c> or <c>~/.config</c> on Linux/Mac.
+    /// </summary>
+    ///
+    /// <returns>
+    /// <c>%APPDATA%/undercut-f1/config.json</c> on Windows,
+    /// <c>$XDG_CONFIG_HOME/undercut-f1/config.json</c> or <c>~/.config/undercut-f1/config.json</c> on Mac/Linux.
+    /// </returns>
+    public static string ConfigFilePath => GetConfigFilePath();
+
     /// <summary>
     /// Prefer to use FFmpeg (<c>ffplay</c>) for audio playback (e.g. Team Radio) instead of more native options
     /// such as <c>mpg123</c> or <c>afplay</c>. FFmpeg is always used on Windows.
@@ -18,6 +33,34 @@ public sealed record Options : LiveTimingOptions
     /// </summary>
     public GraphicsProtocol? ForceGraphicsProtocol { get; set; } = null;
 
-    /// <inheritdoc cref="ExternalPlayerSync.Options" />
-    public ExternalPlayerSync.Options? ExternalPlayerSync { get; set; }
+    /// <inheritdoc cref="ExternalPlayerSync.SyncOptions" />
+    public ExternalPlayerSync.SyncOptions? ExternalPlayerSync { get; set; }
+
+    private static string GetConfigFilePath()
+    {
+        if (OperatingSystem.IsWindows())
+        {
+            return Path.Join(
+                Environment.GetFolderPath(
+                    Environment.SpecialFolder.ApplicationData,
+                    Environment.SpecialFolderOption.Create
+                ),
+                "undercut-f1",
+                "config.json"
+            );
+        }
+        else
+        {
+            var xdgConfigDirectory = Environment.GetEnvironmentVariable("XDG_CONFIG_HOME");
+            if (string.IsNullOrWhiteSpace(xdgConfigDirectory))
+            {
+                xdgConfigDirectory = Path.Join(
+                    Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
+                    ".config"
+                );
+            }
+
+            return Path.Join(xdgConfigDirectory, "undercut-f1", "config.json");
+        }
+    }
 }
